bind recyclerview with adapter
