+++ app/src/main/java/com/iitp/njack/iitp_connect/common/di/modules/DaoModule.java
@@ -9,6 +9,7 @@ import com.iitp.njack.iitp_connect.data.contest.ContestDao;
 import com.iitp.njack.iitp_connect.data.database.IITPConnectDatabase;
 import com.iitp.njack.iitp_connect.data.facebook.FacebookPostDao;
 import com.iitp.njack.iitp_connect.data.user.UserDao;
+import com.iitp.njack.iitp_connect.data.youtube.PlaylistDao;
 
 import javax.inject.Singleton;
 
@@ -45,5 +46,11 @@ public class DaoModule {
     @Singleton @Provides
     public FacebookPostDao providesFacebookPostDao(IITPConnectDatabase iitpConnectDatabase) {
         return iitpConnectDatabase.facebookPostDao();
+    }
+
+    @Singleton
+    @Provides
+    public PlaylistDao providesPlaylistDao(IITPConnectDatabase iitpConnectDatabase) {
+        return iitpConnectDatabase.playlistDao();
     }
 }
