+++ app/src/main/java/com/iitp/njack/iitp_connect/common/di/modules/android/ActivityBuildersModule.java
@@ -12,6 +12,7 @@ import com.iitp.njack.iitp_connect.core.calendar.detail.ContestDetailActivity;
 import com.iitp.njack.iitp_connect.core.calendar.list.CodingCalendarActivity;
 import com.iitp.njack.iitp_connect.core.profile.ProfileActivity;
 import com.iitp.njack.iitp_connect.core.timetable.TimeTableActivity;
+import com.iitp.njack.iitp_connect.core.youtube.playlist.YoutubeActivity;
 
 import dagger.Module;
 import dagger.android.ContributesAndroidInjector;
@@ -53,4 +54,7 @@ public abstract class ActivityBuildersModule {
 
     @ContributesAndroidInjector
     abstract TimeTableActivity contributeTimeTableActivity();
+
+    @ContributesAndroidInjector
+    abstract YoutubeActivity contributeYoutubeActivity();
 }
