+++ app/src/main/java/com/iitp/njack/iitp_connect/core/youtube/CustomViewBindings.java
@@ -1,0 +1,15 @@
+package com.iitp.njack.iitp_connect.core.youtube;
+
+import android.databinding.BindingAdapter;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+
+public class CustomViewBindings {
+    @BindingAdapter("setAdapter")
+    public static void bindRecyclerViewAdapter(RecyclerView recyclerView,RecyclerView.Adapter<?> adapter){
+        recyclerView.setHasFixedSize(true);
+        recyclerView.setLayoutManager(new LinearLayoutManager(recyclerView.getContext()));
+        recyclerView.setAdapter(adapter);
+    }
+
+}
