+++ app/src/main/java/com/iitp/njack/iitp_connect/data/youtube/YoutubePlaylist.java
@@ -25,4 +25,5 @@ public class YoutubePlaylist {
     public String title;
     public String date;
     public Long itemCount;
+    public String imageUrl;
 }
